added styling and navigation to sign in/login page
