+++ client/src/pages/LogIn/LogIn.styles.js
@@ -1,0 +1,7 @@
+import { styled } from "styled-components";
+
+export const SmallContainer = styled.div`
+margin: 10px;
+background-color: white;
+opacity: 0.75; 
+`;
